С учётом замечаний/предложений.
